show top types for different periods
This follows the existing pattern used for period based charts.
Since this isn't a chart but a list, maybe that's not what is intended.
